+++ indoorcycling2/addsensorconfigurationdialog.cpp
@@ -99,6 +99,7 @@ AddSensorConfigurationDialog::AddSensorConfigurationDialog(
 
 AddSensorConfigurationDialog::~AddSensorConfigurationDialog()
 {
+    _antCentralDispatch->closeAllChannels();
     delete _ui;
 }
 
@@ -209,6 +210,7 @@ void AddSensorConfigurationDialog::on_searchSensorsButton_clicked()
     performSearch(indoorcycling::AntSensorType::SPEED_AND_CADENCE);
     performSearch(indoorcycling::AntSensorType::CADENCE);
     performSearch(indoorcycling::AntSensorType::SPEED);
+    performSearch(indoorcycling::AntSensorType::SMART_TRAINER);
 }
 
 void AddSensorConfigurationDialog::fillUsbStickPresentLabel(bool present)
@@ -247,9 +249,13 @@ void AddSensorConfigurationDialog::handleSensorValue(const indoorcycling::Sensor
 
 void AddSensorConfigurationDialog::performSearch(indoorcycling::AntSensorType sensorType)
 {
-    _ui->simulationSettingsGroupBox->setEnabled(false);
     int row = rowForSensorType(sensorType);
     if (row >= 0) {
+        if (!_antCentralDispatch->searchForSensorType(sensorType)) {
+            return;
+        }
+        _ui->simulationSettingsGroupBox->setEnabled(false);
+
         _configurations.remove(sensorType);
         // disable push button
         _ui->searchTableWidget->cellWidget(row, columnNumber(SearchTableColumn::BUTTON))
@@ -263,7 +269,6 @@ void AddSensorConfigurationDialog::performSearch(indoorcycling::AntSensorType se
         bar->setValue(0);
 
         _currentSearches.insert(sensorType);
-        _antCentralDispatch->searchForSensorType(sensorType);
     }
 }