+++ mainlib/ant/antchannelhandler.cpp
@@ -21,9 +21,12 @@
 
 #include <QtCore/QMap>
 #include <QtCore/QtDebug>
+#include <QtCore/QTimer>
 
 namespace {
 const int SEARCH_TIMEOUT = 10; //seconds.
+const int ACKNOWLEDGED_MESSAGE_QUEUE_SIZE = 4; // 1 acknowledged messages.
+
 using indoorcycling::AntChannelHandler;
 const QMap<AntChannelHandler::ChannelState,QString> CHANNEL_STATE_STRINGS(
 {
@@ -50,8 +53,14 @@ AntChannelHandler::AntChannelHandler(const int channelNumber, const AntSensorTyp
 
 void AntChannelHandler::queueAcknowledgedMessage(const AntMessage2 &message)
 {
-    _acknowledgedMessagesToSend.push_back(message);
-    sendNextAcknowledgedMessage();
+    if (_acknowledgedMessagesToSend.size() > ACKNOWLEDGED_MESSAGE_QUEUE_SIZE) {
+        qDebug() << "Queueing Acknowledged message, queue is full. Removing oldest message from queue";
+        _acknowledgedMessagesToSend.pop();
+    }
+    _acknowledgedMessagesToSend.push(message);
+    if (_acknowledgedMessagesToSend.size() > 1) {
+        qDebug() << "Queued Acknowledged message, queue size" << _acknowledgedMessagesToSend.size();
+    }
 }
 
 quint8 AntChannelHandler::channelNumber() const
@@ -74,23 +83,31 @@ void AntChannelHandler::channelOpened()
     // empty.
 }
 
+/**
+ * An acknowledged message transfer has been completed. Remove the message from the queue. If there are any more
+ * messages on the queue, they will be sent after the next broadcast message has been received.
+ */
 void AntChannelHandler::transferTxCompleted()
 {
-    _currentAcknowledgedMessage = AntMessage2();
-    _acknowledgedMessagesToSend.pop_front();
-    sendNextAcknowledgedMessage();
+    _acknowledgedMessageInFlight = false;
+    _acknowledgedMessagesToSend.pop();
+    qDebug() << "dequeued Acknowledged message, queue size" << _acknowledgedMessagesToSend.size();
 }
 
+/**
+ * An acknowledged message transfer has failed. The message will be kept on front of the queue and will be sent after
+ * the next broadcast message has been received.
+ */
 void AntChannelHandler::transferTxFailed()
 {
     qDebug() << "An acknowledged message failed. Trying it again";
-    _currentAcknowledgedMessage = AntMessage2();
+    _acknowledgedMessageInFlight = false;
     sendNextAcknowledgedMessage();
 }
 
 AntChannelHandler::~AntChannelHandler()
 {
-    qDebug() << "Deleting AntChannelHandler for channel type" << static_cast<int>(_sensorType);
+    qDebug() << "Deleting AntChannelHandler for channel type" << ANT_SENSOR_TYPE_STRINGS[_sensorType] << static_cast<int>(_sensorType);
 }
 
 AntSensorType AntChannelHandler::sensorType() const
@@ -136,7 +153,11 @@ void AntChannelHandler::handleChannelEvent(const AntChannelEventMessage &message
         break;
     case AntChannelEventMessage::MessageCode::EVENT_RX_FAILED:
         qDebug() << "RX Failure on channel" << _channelNumber;
-        break;;
+        break;
+    case AntChannelEventMessage::MessageCode::EVENT_RX_FAIL_GO_TO_SEARCH:
+        qDebug() << _channelNumber << "Too many messages missed, going to search mode";
+        setState(ChannelState::SEARCHING);
+        break;
     case AntChannelEventMessage::MessageCode::EVENT_CHANNEL_CLOSED:
         qDebug() << "Channel closed by ANT+ stick.";
         if (_state != ChannelState::UNASSIGNED) {
@@ -160,6 +181,8 @@ void AntChannelHandler::handleChannelEvent(const AntChannelEventMessage &message
 
 void AntChannelHandler::handleBroadcastEvent(const BroadCastMessage &broadcastMessage)
 {
+    // we'll check if there are any acknowledged message to send. If so we'll send one now.
+    sendNextAcknowledgedMessage();
     if (_state == ChannelState::SEARCHING) {
         handleFirstBroadCastMessage(broadcastMessage);
     } else if (_state == ChannelState::TRACKING) {
@@ -199,11 +222,16 @@ void AntChannelHandler::assertMessageId(const AntMessage2::AntMessageId expected
     }
 }
 
+/**
+ * Send thee next Acknowledged message, if there's any to send. This should only be called after a broadcastmessage
+ * has been received.
+ */
 void AntChannelHandler::sendNextAcknowledgedMessage()
 {
-    if (_currentAcknowledgedMessage.isNull() && !_acknowledgedMessagesToSend.empty()) {
-        _currentAcknowledgedMessage = _acknowledgedMessagesToSend.front();
-        emit antMessageGenerated(_currentAcknowledgedMessage);
+    if (!_acknowledgedMessageInFlight && !_acknowledgedMessagesToSend.empty()) {
+        qDebug() << "sending acknowledged message";
+        _acknowledgedMessageInFlight = true;
+        emit antMessageGenerated(_acknowledgedMessagesToSend.front());
     }
 }