+++ mainlib/ant/antchannelhandler.h
@@ -21,7 +21,7 @@
 #define ANTCHANNELHANDLER_H
 
 #include <memory>
-#include <deque>
+#include <queue>
 #include <QtCore/QObject>
 
 #include "antmessage2.h"
@@ -80,6 +80,10 @@ protected:
     explicit AntChannelHandler(const int channelNumber, const AntSensorType sensorType,
                                AntSportPeriod channelPeriod, QObject* parent);
 
+    /**
+     * Queue an Acknowledged message. The message will be sent after all other messages on the queue have been sent.
+     * If the queue is filled up to ACKNOWLEDGED_MESSAGE_QUEUE_CAPACITY, the message will be dropped and false will be returned.
+     */
     void queueAcknowledgedMessage(const AntMessage2 &message);
 
     quint8 channelNumber() const;
@@ -103,23 +107,25 @@ protected:
      * The default implementation is empty.
      */
     virtual void channelOpened();
-
 private:
     void setState(ChannelState state);
     void advanceState(const AntMessage2::AntMessageId messageId);
     void handleFirstBroadCastMessage(const BroadCastMessage&);
     void assertMessageId(const AntMessage2::AntMessageId expected, const AntMessage2::AntMessageId actual);
-    void sendNextAcknowledgedMessage();
     void transferTxCompleted();
     void transferTxFailed();
+    void sendNextAcknowledgedMessage();
 
     const int _channelNumber;
     int _deviceNumber;
     const AntSensorType _sensorType;
     const AntSportPeriod _channelPeriod;
     ChannelState _state;
-    AntMessage2 _currentAcknowledgedMessage;
-    std::deque<AntMessage2> _acknowledgedMessagesToSend;
+
+    /** A queue of acknowledged messages that must be sent. */
+    std::queue<AntMessage2> _acknowledgedMessagesToSend;
+    /** If this is true, there is an acknowledged message in flight and we cannot send a new one. */
+    bool _acknowledgedMessageInFlight = false;
 };
 
 class AntMasterChannelHandler: public AntChannelHandler