+++ mainlib/model/simulation.h
@@ -48,7 +48,6 @@ public:
     QTime runTime() const;
 signals:
     void slopeChanged(float slope);
-    void altitudeChanged(float altitude);
     void runTimeChanged(QTime& runTime);
 
     void playing(bool playing);