+++ test/main.cpp
@@ -1,6 +1,7 @@
 #include "antmessage2test.h"
 #include "profiletest.h"
 #include "reallifevideocachetest.h"
+#include "ridefilewritertest.h"
 #include "rollingaveragecalculatortest.h"
 #include "virtualtrainingfileparsertest.h"
 #include "virtualpowertest.h"
@@ -18,6 +19,8 @@ int main(int, char**) {
     QTest::qExec(&profileTest);
     RollingAverageCalculatorTest racTest;
     QTest::qExec(&racTest);
+    RideFileWriterTest rideFileWriterTest;
+    QTest::qExec(&rideFileWriterTest);
     RealLifeVideoCacheTest saveAndLoadTest;
     QTest::qExec(&saveAndLoadTest);
 }